+ chg: replace tellenc by "Compact Encoding Detection" CED
