+++ src/Notepad3.vcxproj
@@ -99,7 +99,7 @@
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
-      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;..\ced\ced;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <BasicRuntimeChecks>EnableFastChecks</BasicRuntimeChecks>
       <DebugInformationFormat>EditAndContinue</DebugInformationFormat>
       <MinimalRebuild>false</MinimalRebuild>
@@ -154,7 +154,7 @@
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <ClCompile>
-      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;..\ced\ced;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <BasicRuntimeChecks>EnableFastChecks</BasicRuntimeChecks>
       <DebugInformationFormat>ProgramDatabase</DebugInformationFormat>
       <MinimalRebuild>false</MinimalRebuild>
@@ -208,7 +208,7 @@
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
-      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;..\ced\ced;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <MultiProcessorCompilation>true</MultiProcessorCompilation>
       <Optimization>MaxSpeed</Optimization>
       <PreprocessorDefinitions>WIN32;STATIC_BUILD;SCI_LEXER;NDEBUG;%(PreprocessorDefinitions)</PreprocessorDefinitions>
@@ -264,7 +264,7 @@
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <ClCompile>
-      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>..\scintilla\include;..\scintilla\lexlib;..\scintilla\src;..\ced\ced;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <MultiProcessorCompilation>true</MultiProcessorCompilation>
       <Optimization>MaxSpeed</Optimization>
       <PreprocessorDefinitions>_WIN64;STATIC_BUILD;SCI_LEXER;NDEBUG;%(PreprocessorDefinitions)</PreprocessorDefinitions>
@@ -318,6 +318,10 @@
     </ManifestResourceCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\ced\ced\compact_enc_det\compact_enc_det.cc" />
+    <ClCompile Include="..\ced\ced\compact_enc_det\compact_enc_det_hint_code.cc" />
+    <ClCompile Include="..\ced\ced\util\encodings\encodings.cc" />
+    <ClCompile Include="..\ced\ced\util\languages\languages.cc" />
     <ClCompile Include="..\crypto\crypto.c" />
     <ClCompile Include="..\crypto\rijndael-alg-fst.c" />
     <ClCompile Include="..\crypto\rijndael-api-fst.c" />
@@ -326,12 +330,28 @@
     <ClCompile Include="Dlapi.c" />
     <ClCompile Include="Edit.c" />
     <ClCompile Include="Encoding.c" />
+    <ClCompile Include="EncodingCED.cpp" />
     <ClCompile Include="Helpers.c" />
     <ClCompile Include="Notepad3.c" />
     <ClCompile Include="Print.cpp" />
     <ClCompile Include="Styles.c" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\ced\ced\compact_enc_det\compact_enc_det.h" />
+    <ClInclude Include="..\ced\ced\compact_enc_det\compact_enc_det_generated_tables.h" />
+    <ClInclude Include="..\ced\ced\compact_enc_det\compact_enc_det_generated_tables2.h" />
+    <ClInclude Include="..\ced\ced\compact_enc_det\compact_enc_det_hint_code.h" />
+    <ClInclude Include="..\ced\ced\util\basictypes.h" />
+    <ClInclude Include="..\ced\ced\util\case_insensitive_hash.h" />
+    <ClInclude Include="..\ced\ced\util\commandlineflags.h" />
+    <ClInclude Include="..\ced\ced\util\encodings\encodings.h" />
+    <ClInclude Include="..\ced\ced\util\encodings\encodings.pb.h" />
+    <ClInclude Include="..\ced\ced\util\languages\languages.h" />
+    <ClInclude Include="..\ced\ced\util\languages\languages.pb.h" />
+    <ClInclude Include="..\ced\ced\util\logging.h" />
+    <ClInclude Include="..\ced\ced\util\port.h" />
+    <ClInclude Include="..\ced\ced\util\string_util.h" />
+    <ClInclude Include="..\ced\ced\util\varsetter.h" />
     <ClInclude Include="..\crypto\crypto.h" />
     <ClInclude Include="..\crypto\rijndael-alg-fst.h" />
     <ClInclude Include="..\crypto\rijndael-api-fst.h" />
@@ -340,7 +360,6 @@
     <ClInclude Include="Dlapi.h" />
     <ClInclude Include="Edit.h" />
     <ClInclude Include="Encoding.h" />
-    <ClInclude Include="EncTables.hpp" />
     <ClInclude Include="Helpers.h" />
     <ClInclude Include="Notepad3.h" />
     <ClInclude Include="SciCall.h" />
